requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.11.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==26.1.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
billiard==4.2.4
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.6.3
    # via -r requirements/base.txt
certifi==2026.2.25
    # via
    #   -r requirements/base.txt
    #   requests
cffi==2.0.0
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.4.6
    # via
    #   -r requirements/base.txt
    #   requests
click==8.3.1
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
    #   import-linter
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1.2
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==3.0.0
    # via -r requirements/test.in
coverage[toml]==7.13.5
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==46.0.7
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-debug-toolbar
    #   django-model-utils
    #   django-simple-history
    #   django-stubs
    #   django-stubs-ext
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-organizations
    #   openedx-events
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-debug-toolbar==6.2.0
    # via -r requirements/test.in
django-model-utils==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
django-simple-history==3.11.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
django-stubs==6.0.1
    # via
    #   -r requirements/test.in
    #   djangorestframework-stubs
django-stubs-ext==6.0.1
    # via django-stubs
django-waffle==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.17.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-organizations
djangorestframework-stubs==3.16.8
    # via -r requirements/test.in
dnspython==2.8.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-ccx-keys==2.0.2
    # via
    #   -r requirements/base.txt
    #   openedx-events
edx-django-utils==8.0.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   openedx-events
edx-drf-extensions==10.6.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
edx-opaque-keys[django]==4.0.0
    # via
    #   -r requirements/base.txt
    #   edx-ccx-keys
    #   edx-drf-extensions
    #   edx-organizations
    #   openedx-events
edx-organizations==8.0.0
    # via -r requirements/base.txt
fastavro==1.12.1
    # via
    #   -r requirements/base.txt
    #   openedx-events
freezegun==1.5.5
    # via -r requirements/test.in
grimp==3.14
    # via import-linter
idna==3.11
    # via
    #   -r requirements/base.txt
    #   requests
import-linter==2.11
    # via -r requirements/test.in
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
kombu==5.6.2
    # via
    #   -r requirements/base.txt
    #   celery
librt==0.9.0
    # via mypy
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mock==5.2.0
    # via -r requirements/test.in
mypy==1.19.1
    # via -r requirements/test.in
mypy-extensions==1.1.0
    # via mypy
mysqlclient==2.2.8
    # via -r requirements/test.in
openedx-events==11.1.0
    # via -r requirements/base.txt
packaging==26.0
    # via
    #   -r requirements/base.txt
    #   kombu
    #   pytest
pathspec==1.0.4
    # via mypy
pillow==12.1.1
    # via
    #   -r requirements/base.txt
    #   edx-organizations
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
prompt-toolkit==3.0.52
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==7.2.2
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==3.0
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.20.0
    # via
    #   pytest
    #   rich
pyjwt[crypto]==2.12.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.16.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.6.2
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==9.0.3
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==7.1.0
    # via -r requirements/test.in
pytest-django==4.12.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
    #   freezegun
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.3
    # via code-annotations
requests==2.33.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
rich==15.0.0
    # via import-linter
rules==3.5
    # via -r requirements/base.txt
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   edx-ccx-keys
    #   python-dateutil
sqlparse==0.5.5
    # via
    #   -r requirements/base.txt
    #   django
    #   django-debug-toolbar
stevedore==5.7.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomlkit==0.14.0
    # via -r requirements/base.txt
types-pyyaml==6.0.12.20260408
    # via
    #   django-stubs
    #   djangorestframework-stubs
typing-extensions==4.15.0
    # via
    #   -r requirements/base.txt
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   edx-opaque-keys
    #   grimp
    #   import-linter
    #   mypy
tzdata==2025.3
    # via
    #   -r requirements/base.txt
    #   kombu
tzlocal==5.3.1
    # via
    #   -r requirements/base.txt
    #   celery
urllib3==2.6.3
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
